These proposal builder functions don't need to be hooks or created via hooks
